requirements.txt
asgiref
attrs
cached-property
cachetools
certifi
charset-normalizer
cycler
Django
django-storages
djangorestframework
idna
isodate
kiwisolver
lxml
numpy
packaging
pandas
patsy
Pillow
platformdirs
plotly
protobuf
psycopg2
pyasn1
pyasn1-modules
pyparsing
python-dateutil
pytz
requests
requests-file
requests-toolbelt
rsa
scipy
six
sqlparse
tenacity
urllib3
zeep
django-crispy-forms
stringdist
